Move FutureTimepoint from network/generics to core
